split chaindata into multiple files
